Move non-animation Mesh code to MeshData class
