+++ Client/Client.vcxproj.filters
@@ -108,6 +108,9 @@
     <ClCompile Include="Renderer\Material.cpp">
       <Filter>Source Files\Renderer</Filter>
     </ClCompile>
+    <ClCompile Include="Renderer\MeshData.cpp">
+      <Filter>Source Files\Renderer</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
@@ -177,6 +180,9 @@
       <Filter>Header Files\Sound</Filter>
     </ClInclude>
     <ClInclude Include="Renderer\Material.h">
+      <Filter>Header Files\Renderer</Filter>
+    </ClInclude>
+    <ClInclude Include="Renderer\MeshData.h">
       <Filter>Header Files\Renderer</Filter>
     </ClInclude>
   </ItemGroup>
