+++ Client/Renderer/Mesh.h
@@ -7,28 +7,7 @@
 #include <assimp/scene.h>
 #include <assimp/postprocess.h>
 #include <unordered_map>
-
-// The maximum number of bones that can affect a vertex.
-constexpr auto BONES_PER_VERTEX = 4;
-
-// Type of number used for indices in the element buffer.
-typedef unsigned int ElementIndex;
-
-// Container for the position vector and normal vector of a vertex.
-struct Vertex {
-	vec3 pos;
-	vec3 normal;
-	vec3 tangent = vec3(0.0f);
-	vec3 bitangent = vec3(0.0f);
-	vec2 texCoords;
-	GLuint bones[BONES_PER_VERTEX] = { 0 };
-	GLfloat weights[BONES_PER_VERTEX] = { 0.0f };
-};
-
-struct Bone {
-	std::string name;
-	mat4 offset;
-};
+#include "MeshData.h"
 
 // The Mesh class is used to draw a collection of vertices using triangles in the
 // order given by indices.
@@ -36,32 +15,15 @@ class Mesh {
 	friend class Model;
 
 	private:
-	GLuint VAO, VBO, EBO;
-	GLsizei numVertices;
-	std::vector<Bone> bones;
-	std::unordered_map<std::string, int> boneIds;
-	std::unordered_map<const char*, aiNodeAnim> nodeAnims;
 	std::vector<mat4> boneTransformations;
-	aiMesh *mesh;
-	const aiScene *scene;
+	MeshData *data;
 	int animationId = -1;
-
-	// Gets the vertices of the aiMesh into a list of Vertex instances.
-	static std::vector<Vertex> loadVertices(aiMesh *mesh);
-
-	// Gets the indices for the vertices into a list of numbers.
-	static std::vector<ElementIndex> loadIndices(aiMesh *mesh);
-
-	// Sets up which bones will affect which vertices for the model.
-	void loadBones(std::vector<Vertex> &vertices);
 
 	// Sets up the vertex transformations for the current animation.
 	void buildBoneTransformations(float time, aiNode *node, const mat4 &parentTransform);
-	aiNodeAnim *getNodeAnim(const aiAnimation *animation, const std::string &nodeName);
 
 	public:
-	Mesh(aiMesh *mesh, const aiScene *scene);
-	~Mesh();
+	Mesh(MeshData *data);
 
 	void updateAnimation(float time);
